+++ frontend/src/router.js
@@ -2,7 +2,7 @@ import { createRouter, createWebHistory } from 'vue-router'
 import Home from './views/Home.vue'
 import Predict from './views/Predict.vue'
 import Statistics from './views/Statistics.vue'
-import Viewstat from './components/Viewstat.vue'
+import Viewstat from './views/Viewstat.vue'
 
 const routes = [
     {
