+++ frontend/src/views/Viewstat.vue
